# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Management System API",
        "description": "A comprehensive car rental management system with user authentication, vehicle management, booking system, payment processing, and review functionality. This system supports both car rentals and property (home) rentals.",
        "contact": {
            "name": "Car Rental System Support",
            "email": "support@carrentalsystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Car Rental API Server"
        }
    ],
    "paths": {
...
        "/api/chapa/callback": {
            "get": {
                "tags": [
                    "Payment Processing"
                ],
                "summary": "Handle Chapa payment callback for bookings",
                "description": "Webhook endpoint to handle payment verification callbacks from Chapa for booking payments",
                "operationId": "77b3607b24a53d938b425ae3c3143c0b",
                "parameters": [
                    {
                        "name": "tx_ref",
                        "in": "query",
                        "description": "Transaction reference from Chapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "TX-abc123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment verified and processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Payment successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Payment verification failed or missing transaction reference",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Transaction reference is missing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Payment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/google": {
            "get": {
                "tags": [
                    "SSO Authentication"
                ],
                "summary": "Redirect to Google OAuth",
                "description": "Initiate Google OAuth authentication flow",
                "operationId": "0fb6b0e301610e77a2f685bf10e4673e",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "query",
                        "description": "Platform type for appropriate callback",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "web",
                                "mobile"
                            ]
                        },
                        "example": "web"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth page"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}